Import BBS archive with migrate-repo command (#582)
Closes #528!
This PR updates the placeholder migrate-repo command for bbs2gh with real logic to ingest an archive URL and kick off a migration with Octoshift.
A few notes about this PR:

There are some parameters that pass strings like "not-used" to satisfy required parameters in GraphQL and Octoshift.  When github/octoshift#5350 is shipped, these "filler" strings can be removed.
CreateBbsMigrationSource was added to GithubApi.  This is consumed by the bbs2gh MigrateRepoCommand to kick off migrations over GraphQL.
bbs2gh.csproj was added as a reference to OctoshiftCLI.Tests.csproj.  This is necessary to test classes in the OctoshiftCLI.BbsToGithub.Commands namespace.
I opted to use a MigrateRepoCommandArgs class in bbs2gh's MigrateRepoCommand because I felt that it was cleaner and easier to read.
Some unused ADO logic was pulled from bbs2gh's EnvironmentVariableProvider.
As for test coverage, I opted to find parity in the coverage for the GEI tests that is relevant to bbs2gh.  I'm happy to add more coverage if desired 👍

Tasks

 Did you write/update appropriate tests
 Release notes updated (if appropriate)
 Appropriate logging output
 Issue linked
 Docs updated (or issue created) (added bullet point to #586)