Add BBS client/API plumbing (#583)
The first step in migrating BBS repos is to tell the source instance to generate the archive, then to query the REST API to get the state of the export (until it's done). Additionally, we want to be able to query the server to get its version, to make sure it's one that supports the export API. This PR adds a minimal, naïve implementation of `BbsClient` (which abstracts away talking to the BBS source instance) and `BbsApi`, which formalizes the API endpoints we want to hit as we kick off a migration. 

Closes https://github.com/github/gh-gei/issues/524

- [x] Did you write/update appropriate tests
- [x] Appropriate logging output
- [x] Issue linked
- [ ] ~~Release notes updated (if appropriate)~~
- [ ] ~~Docs updated (or issue created)~~

<!--
For docs we should review the docs at:
https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is sufficient to create an issue and link to it here. The doc update can be made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 
https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->