refactor LogAndValidateOptions (#595)
<!--
For the checkboxes below you must check each one to indicate that you either did the relevant task, or considered it and decided there was nothing that needed doing
-->

In `gei migrate-repo` the function `LogAndValidateOptions()` had gotten so large that `dotnet format` was starting to complain if you added anything new to it. Like in this other PR: https://github.com/github/gh-gei/pull/589

I refactored it into separate `LogOptions()` and `ValidateOptions()` functions, which I think is better even if our linter didn't complain.

This is pre-req work for the lock-repo PR and issue #385 

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:
https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is sufficient to create an issue and link to it here. The doc update can be made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 
https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->