Add --lock-source-repo option. (#589)
<!--
For the checkboxes below you must check each one to indicate that you either did the relevant task, or considered it and decided there was nothing that needed doing
-->
This change addresses #385 by adding the cli option `--lock-source-repo`. 

For GHES source, this uses the `lock_repositories` option [as documented here](https://docs.github.com/en/rest/migrations/orgs#start-an-organization-migration).

For GHEC source, this uses the new lockSource option in the GraphAPI [mentioned here](https://github.com/github/migration-friction/issues/618#issuecomment-1221009872).

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created) #597 

<!--
For docs we should review the docs at:
https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is sufficient to create an issue and link to it here. The doc update can be made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 
https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

Co-authored-by: Dylan Smith <dylan-smith@github.com>
Co-authored-by: Dylan Smith <dylanfromwinnipeg@gmail.com>