Add archive generation to `bbs2gh migrate-repo` (#602)
This PR adds a path to `bbs2gh migrate-repo` to generate an archive on a BBS instance if the `--bbs-server-url` argument is present. Because we're making comparatively fewer requests to a BBS instance (compared to GitHub or ADO), the logic isn't quite as complex or robust, but it does generate the archive. 

Closes https://github.com/github/gh-gei/issues/524 (for real this time)

---

- [x] Did you write/update appropriate tests
- [ ] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [ ] Issue linked
- [ ] Docs updated (or issue created)

<!--
For docs we should review the docs at:
https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is sufficient to create an issue and link to it here. The doc update can be made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 
https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

Co-authored-by: Arin Ghazarian <aringhazarian@github.com>
Co-authored-by: Dylan Smith <dylan-smith@github.com>