warn users  not running ado2gh as a  gh extension (#628)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Checking if the user is running ado2gh via a GH CLI extension or as a
stand-alone executable. Doing the check by looking at the path the
executable lives in, if it's being run as a GH CLI extension it should
have a specific path it lives under:
```
Windows: AppContext.BaseDirectory:C:\Users\Dylan\AppData\Local\GitHub CLI\extensions\gh-gei\
Linux: AppContext.BaseDirectory:/home/dylan/.local/share/gh/extensions/gh-gei/
MacOS: AppContext.BaseDirectory:/Users/iomekam/.local/share/gh/extensions/gh-gei/
```

If they are NOT using it via GH extension we'll spit out a warning
recommending that they switch.

Closes #584 

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->