removed --ssh option entirely (#631)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Closes #270 

On March 28 [v0.10](https://github.com/github/gh-gei/releases/tag/v0.10)
we deprecated the `--ssh` flag, but we left in the CLI as a hidden
option that did nothing other than warn users not to use it anymore. We
didn't fully remove it because we didn't want to break customers scripts
that had already been generated with `--ssh` in there.

That was 6 months ago, and now it's time to remove it entirely from our
codebase.

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->