Use `X-RateLimit-Reset` headers to throttle GitHub requests (#623)
This PR adds code to `GithubClient` to introspect the
`X-RateLimit-Remaining` and `X-RateLimit-Reset` headers on GitHub
requests (both REST and GraphQL; I checked) to pause the CLI until their
rate limit resets. On each request, it checks if the
`X-RateLimit-Remaining` value is `0` (indicating no requests left), then
extracts the `X-RateLimit-Reset` header (which is the Unix epoch second
that the rate limit resets), calculates the diff between that and the
present time, and sets that as `_retryDelay`. Once `_retryDelay` is set,
`GithubClient` behaves identically to `AdoClient`.

Closes https://github.com/github/gh-gei/issues/494
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [ ] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [ ] Issue linked
- [ ] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

Co-authored-by: Arin Ghazarian <aringhazarian@github.com>
Co-authored-by: Dylan Smith <dylan-smith@github.com>