Introduced CommandArgs classes for all gei and Octoshift commands (#643)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->
Refactored all remaining commands in Octoshift/gei/ado2gh to use Args
classes. Didn't touch bbs2gh yet.

For gei commands that inherit from an Octoshift shared command class
then change an arg name, we had duplicate the Args class, and do a bit
of mapping code in the sub-classed command. This bit of cruft will go
away with the last step of this refactor when we change how we map
command-line options to the Args class using a custom binder.

Part of #637 

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->