CommandLine plumbing infrastructure (#652)
Related to #639

This is all the required infrastructure changes in form of some
extension method to do the plumbing for the new CommandLine. Here is a
[sample
branch](https://github.com/github/gh-gei/tree/upgrade-bbs2gh-to-new-command-line)
that applies these changes to `bbs2gh`.
I went ahead and simplified it even more by creating an extension to add
all commands to the `rootCommand`. Here is how it's supposed to be used:

```c#
public static async Task<int> Main(string[] args) // note that Main now returns a Task<int> instead of Task
{
    Logger.LogDebug("Execution Started");
 
    var serviceCollection = new ServiceCollection();
    ... // service registrations
    var serviceProvider = serviceCollection.BuildServiceProvider();

    var rootCommand = new RootCommand("Automate end-to-end Bitbucket Server to GitHub migrations.")
          .AddCommands(serviceProvider);

    ...
}
```

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

Co-authored-by: Dylan Smith <dylan-smith@github.com>
Co-authored-by: Dylan Smith <dylanfromwinnipeg@gmail.com>