Move GHES archive-downloading logic out of `AzureApi` (#679)
When we shipped #653, we missed a case where omitting
`--azure-storage-connection-string` causes the app to crash, because
`AzureApiFactory.Create` was running anyway. After a bit of digging, I
discovered that we were using `AzureApi` to make unauthenticated HTTP
GETs to download the archives off of the GHES instance. This PR moves
`AzureApi.DownloadArchive` into its own class, so that we can
encapsulate making unauthenticated HTTP GETs.


- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [ ] Issue linked
- [ ] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->