Add Kerberos support (#684)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Issue #678 

Added support for kerberos authentication when calling BBS API's. To do
this all we need to do is configure our HttpClient to have
`UseDefaultCredentials = true` and NOT set the normal Authorization
header that we set. This needs to be set at the time we construct the
HttpClient. Similar to how we did the NoSSL support, I added a couple
different named HttpClient's into our DI container ("Default" and
"Kerberos") configured appropriately. Then I added a new
`CreateKerberos()` function to `BbsApiFactory` that doesn't require a
username/password. Finally I added a 2nd constructor to `BbsApi` that
doesn't require a username/password and doesn't attempt to set the
Authorization header.

I added a `--kerberos` flag to both `bbs2gh generate-script` and `bbs2gh
migrate-repo`. For now this is a hidden option.

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->