Output information about the progress of org-to-org migrations as repos are migrated
When waiting for an organization-to-organization migration with the
`wait-for-migration` command, the only feedback that the CLI provides
on the progress of the migration is its `state`.

Org-to-org migrations potentially stay in the `REPO_MIGRATION` `state`
for a long time if there are many repos to be migrated, and this can be
a worrying time - as a user, you are likely to wonder "is my migration
actually moving forward"?.

The `node` query for `OrganizationMigration`s allows you to fetch the
`remainingRepositoriesCount` and `totalRepositoriesCount` for the
migration.

This commit starts grabbing that data and logging a progress indicator
(`1/10 repos migrated`) to the console.

Note that the GraphQL API currently returns `0` for the two count
fields if the number of repos has not yet been completed. There is an
outstanding bug to fix this and return `null` instead. This code is
future-proof, able to handle both values.

Fixes https://github.com/github/gh-gei/issues/699.