Add retry logic to GQL get api calls (#686)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Closes #668 

Our normal retry on failed GET's logic doesn't apply to GQL queries,
because technically they aren't doing a HTTP GET but rather a POST to
the GQL endpoint, so our retry logic doesn't kick in. They also aren't
always getting a normal HTTP error, instead they often are failing with
a HTTP 200 but with a GQL error message in the response body.

There are 9 places in `GithubApi` where we do a read-only query via GQL.
2 of those places are actually dead code that isn't used anywhere and I
deleted it:
- `GetOrganizationMigrationState`
- `GetMigrationStates`

6 instances where I updated the code to inspect the GQL response body
and check for any errors, and if so apply our normal retry logic:
- `GetOrganizationId`
- `GetEnterpriseId`
- `GetOrganizationMigration`
- `GetMigration`
- `GetMigrationLogUrl`
- `GetMannequins`

Then there was 1 instance where doing the retry logic will be a bit
trickier - `GetUserId` - because it assumes a GQL error means the user
doesn't exist and has some specific behavior in that scenario that
relies on it returning null instead of an exception. For this one I
didn't tackle it in this PR and created a new issue for us to tackle it
at some point in the future: #687

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->