Output information about the progress of org-to-org migrations as repos are migrated (#700)
When waiting for an organization-to-organization migration with the
`wait-for-migration` command, the only feedback that the CLI provides on
the progress of the migration is its `state`.

Org-to-org migrations potentially stay in the `REPO_MIGRATION` `state`
for a long time if there are many repos to be migrated, and this can be
a worrying time - as a user, you are likely to wonder "is my migration
actually moving forward"?.

The `node` query for `OrganizationMigration`s allows you to fetch the
`remainingRepositoriesCount` and `totalRepositoriesCount` for the
migration.

This commit starts grabbing that data and logging a progress indicator
(`1/10 repos migrated`) to the console.

Note that the GraphQL API currently returns `0` for the two count fields
if the number of repos has not yet been completed. There is an
outstanding bug to fix this and return `null` instead. This code is
future-proof, able to handle both values.

Fixes https://github.com/github/gh-gei/issues/699.

_I don't intend to make it a habit to make code changes, but I wanted to
see this enhancement added before org-to-org ships, and I wanted to get
familiar with the codebase 😊 _

<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)