BBS Integration tests (#696)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Stood up a BBS server in our shared Azure Subscription
(e2e-bbs-8-5-0-linux-2204.eastus.cloudapp.azure.com). It's running the
latest version of BBS (8.5.0) and Ubuntu 22.04.

Manually created 2 projects and 2 repos in BBS (in the future we may
automate this like we have for ADO/GH), and added a new integration test
to migrate BBS->GH.

Included asserts to check if the GH repo exists and is initialized.

The default SSH cert that Azure generates didn't seem to work when using
Ubuntu 22.04 (didn't work with our SSH.Net nuget library, but did work
when directly using SSH in a terminal). I found a [workaround in an
issue](https://github.com/sshnet/SSH.NET/issues/825#issuecomment-1139440419)
from the SSH.Net repo that I applied in this PR. I don't fully
understand exactly what it's doing, but it looks like it's updating the
signature on the key in-memory before using it. Whatever it's doing, it
makes keys that previously didn't work now work. So it seems like a good
thing to me.

Closes #553 

Future work:
- automate the creation of BBS test data (like we do for ADO/GH)
- Include downloading the migration logs (and asserting that they exist)
- Add BBS windows server (once we support SMB)
- Add AWS
- Add another BBS server with the oldest version we support (5.14 I
think)

Checklist
- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->