Adding migration of GitHub Secret Scanning Alerts (#592)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created) (#715 )

This is the first pass on the requirements to migrate secret scanning
remediations between two repositories, provides the functionality for
#536.

This is being opened to start a code review and then complete the
support of this command line and service. There are some necessary
manual steps required as part of the larger migration in that the repo
is migrated and secret scanning has been reactivated (and the backfill
scan completed) before this command line makes sense to execute.

Created a new issue to add an integration test to cover this
functionality in the future: #717

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

Co-authored-by: Peter Murray <peter-murray@users.noreply.github.com>
Co-authored-by: Arin Ghazarian <aringhazarian@github.com>
Co-authored-by: Dylan Smith <dylanfromwinnipeg@gmail.com>