Add support to `bbs2gh migrate-repo` to only generate the archive on the BBS server (#723)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Closes #729 

This adds support for calling `bbs2gh migrate-repo` with a combination
of args that will cause it to tell BBS to generate the archive and do
nothing else.

This can be useful in a few scenarios, but for right now the goal is to
support a customer that uses different proxy servers depending on if
they are talking to an internal server (i.e. BBS) or an external server
(i.e. Azure/GitHub). With this change you will be able to call
`migrate-repo` once to trigger the BBS archive generation, then switch
up your proxy settings and call `migrate-repo` a 2nd time to tell it to
upload to azure and call github.com.

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

Co-authored-by: Arin Ghazarian <aringhazarian@github.com>