Added BBS environment reset code for e2e tests (#733)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Added code to do a similar thing we do with our other e2e tests, where
the test is responsible for resetting the test data that it expects in
the test environment. This makes it easy to stand up new environments,
and prevent accidentally breaking test environments by unintentional
manual changes in the BBS server(s). It also makes it easy to add new
data shapes to our test environments, as we just add them in one place
(in the code) rather than having to manually touch each of our BBS test
servers (and we will have several of them before we're done with the BBS
work).

Right now these additions will check if the BBS project exists, and if
so iterate and delete all repos in the project, then delete the project.
Then create the project(s) and repo(s) from scratch.

It uses libgit2sharp to initialize the repos. This dependency only
exists in our e2e test code, not in the CLI that we ship to customers.
libgit2sharp comes with native binaries for various linux distro's, but
only includes ubuntu 18.04 but our build servers use ubuntu 20.04 (or
maybe 22.04). The workaround is to use the LD_LIBRARY_PATH env var when
building our integration test project to force it to link to the ubuntu
18.04 binary
(https://github.com/libgit2/libgit2sharp/issues/1703#issuecomment-619497921)

This has pretty much everything we need to support testing against BBS
5.14 also (#570 ) except for some reason the initialize repo code fails
consistently when run against BBS 5.14. Once somebody figures out why
that fails, you just need to add an extra [InlineData] attribute to the
test case and it will run for both.

Closes #573 

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->