Update log message after blob storage upload to not refer to a specific blob storage provider
When running an indirect migration, after finishing the upload of
the archives to blob storage, the CLI outputs the following line
to the logs:

> Archives uploaded to Azure Blob Storage, now starting migration...

It's great to tell the user that the blob storage upload is done,
but we now support Amazon S3 as well as Azure Blob Storage, but
this log line *always* mentions Azure.

This updates the log output to just refer to "blob storage" rather
than a specific provider. Previous lines in the logs will make it
clear which provider was used, e.g.:

> Uploading archive 2022-11-16_22-24-14-10-metadata_archive.tar.gz to AWS S3

Fixes #740.