Allow blob credentials to be optional after GHES 3.8 (#771)
Fixes #755 

I didn't want to move around too much logic so just pulled out the GHES
version check to happen earlier to we can properly validate args and
then pass that along to the migrations API logic so we don't call the
meta API twice.
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [ ] Did you write/update appropriate tests
- [ ] Release notes updated (if appropriate)
- [ ] Appropriate logging output
- [ ] Issue linked
- [ ] Docs updated - created https://github.com/github/gh-gei/issues/770

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

Functionally tested as shown below. Real integration tests will be added
in https://github.com/github/gh-gei/issues/756.

#### Tested BAU behavior with older GHES version (3.6.0)

![image](https://user-images.githubusercontent.com/2413532/211442096-98a3439d-f5a7-4d33-b22e-860a7b9566b4.png)

#### Tested new behavior with new GHES version (3.8.0) - no credentials
set in UI

![image](https://user-images.githubusercontent.com/2413532/211897540-bfcf7ae8-b6e0-40b3-98e2-1dc0e18e3585.png)

Error message in cli (moved out of verbose log) since this is an
expected error we should surface.

#### Tested new behavior with new GHES version (3.8.0) - credentials set
in UI

![image](https://user-images.githubusercontent.com/2413532/211442344-1d3dc5ac-576d-480c-8453-db388817310f.png)

![image](https://user-images.githubusercontent.com/2413532/211442379-7ebdfd92-199a-4e16-8b08-f38a3129e123.png)

#### Tested new behavior with new GHES version (3.8.0) - credentials set
in UI and in cli
<img width="1904" alt="image"
src="https://user-images.githubusercontent.com/2413532/211442465-dffe5b79-bf64-4942-ad82-ac92fa1ac98f.png">
Can see the note that the user does not need to set in cli anymore.

Co-authored-by: Dylan Smith <dylanfromwinnipeg@gmail.com>